requirements.txt
torch
torchvision
numpy
uvicorn
fastapi
Pillow
python-multipart